+++ AlchemistCode/Potions/Soporific.cs
@@ -1,0 +1,27 @@
+using System.Collections.Generic;
+using MegaCrit.Sts2.Core.Commands;
+using MegaCrit.Sts2.Core.Entities.Cards;
+using MegaCrit.Sts2.Core.Entities.Creatures;
+using MegaCrit.Sts2.Core.Entities.Potions;
+using MegaCrit.Sts2.Core.GameActions.Multiplayer;
+using MegaCrit.Sts2.Core.HoverTips;
+using MegaCrit.Sts2.Core.Models;
+
+namespace Alchemist.AlchemistCode.Potions;
+
+public class Soporific : AlchemistPotion, IBrewOnly
+{
+    // Event rarity keeps this out of every rarity-filtered roll and files it under "Special" in the
+    // potion lab, which is where a potion that is "created by other means" belongs
+    public override PotionRarity Rarity => PotionRarity.Event;
+    public override PotionUsage Usage => PotionUsage.CombatOnly;
+    public override TargetType TargetType => TargetType.AnyEnemy;
+
+    public override IEnumerable<IHoverTip> ExtraHoverTips => [HoverTipFactory.Static(StaticHoverTip.Stun)];
+
+    protected override async Task OnUse(PlayerChoiceContext choiceContext, Creature? target)
+    {
+        PotionModel.AssertValidForTargetedPotion(target);
+        await CreatureCmd.Stun(target!);
+    }
+}